Check in install-sh as synced from automake (#110368)
As pointed out in sr #110368, since install-sh is now being installed
as part of autoconf, we should make sure to ship the latest version
rather than the version shipped by the automake that was used to
bootstrap the autoconf release tarball.

The build-aux/fetch.pl script is already supposed to fetch the latest
version, but install-sh is listed in .gitignore so any updates are
discarded when starting from a clean tree.  Correct this.

At the same time, since mdate-sh is *not* installed by autoconf nor is
it directly referenced in any code maintained in the autoconf
repository, remove it from the list of files to fetch and keep it in
.gitignore.

This change exposed a bug in fetch.pl where it would crash when
there was no old copy of a file being updated.

 * .gitignore: Remove /build-aux/install-sh.
 * build-aux/fetch.pl (%to_fetch): Remove build-aux/mdate-sh.
   (slurp): Don’t die on ENOENT, return undef.
   (replace_if_change): Handle $oldcontents being undef.
 * build-aux/install-sh: Is now checked in.