Don’t issue obsoletion warnings for AC_LANG_SAVE/RESTORE (#110375)
The most recently released version of libtool.m4 is five years old as
of this commit, and no new release is likely to appear anytime soon.
It uses AC_LANG_SAVE and AC_LANG_RESTORE, in a way that doesn’t
obviously translate to AC_LANG_PUSH and AC_LANG_POP.  This will need
to be fixed by libtool upstream.  Until that actually happens, disable
the -Wobsolete warnings for AC_LANG_SAVE and AC_LANG_RESTORE.  (They
are still documented as obsolete in the manual, as they have been for
many years.)

Fixes bug #110375.

* lib/autoconf/lang.m4 (AC_LANG_SAVE, AC_LANG_RESTORE): Define with
  AC_DEFUN, not AU_DEFUN; remove manual -Wobsolete warnings.