AC_FUNC_SETPGRP: Don’t depend on the return type of setpgrp.
AC_FUNC_SETPGRP determines whether you have the historic BSD setpgrp,
which takes two arguments and returns int, or the historic POSIX
setpgrp, which takes no arguments and returns int.  Solaris has yet a
third variant, which takes no arguments and returns a pid_t (the new
process group ID).  This difference causes AC_FUNC_SETPGRP’s test
program to fail to compile under AC_LANG([C++]), which in turn causes
the macro to report that setpgrp does take arguments, which is wrong.

It is not worth adding a new result #define for this variant,
since *all* forms of setpgrp are deprecated in favor of setpgid, which
is old enough that it can be used unconditionally.  However, it is
worth documenting that this variant exists, and fixing AC_FUNC_SETPGRP
to produce the right value for its existing result #define on Solaris
with C++.

* lib/autoconf/functions.m4 (AC_FUNC_SETPGRP): Redesign test program to
  not depend on the return type of setpgrp.
* doc/autoconf.texi (AC_FUNC_SETPGRP): Mention that the macro does not
  check for the Solaris variant of setpgrp that returns pid_t.  Change
  programming advice to recommend use of setpgid.