Disentangle HAVE__BOOL from ac_cv_header_stdbool_h.
AC_CHECK_HEADER_STDBOOL is documented to make two checks: whether the
C99 header <stdbool.h> is available and fulfills its
specification (i.e. including it makes the type ‘bool’ and the
constants ‘true’ and ‘false’ available), and, independently, whether
the type ‘_Bool’ is available.

In C++, the type ‘_Bool’ is usually _not_ available, but <stdbool.h>
is still supposed to be include-able and the type ‘bool’ and the
constants ‘true’ and ‘false’ are still supposed to be available
(unconditionally).  However, the test for <stdbool.h> fulfilling its
specification freely used _Bool, and would therefore fail spuriously.
Correct this by checking for _Bool first, and then refactoring the
test program for <stdbool.h> so that it does all its tests using bool,
then repeats them with _Bool only when available.

* lib/autoconf/headers.m4 (AC_CHECK_HEADER_STDBOOL): Do the test for
  _Bool before the test for stdbool.h.  Test semantics of bool
  unconditionally; test _Bool only when HAVE__BOOL is defined.