Create disconnected branch for CI scripts.
I don’t know if this will be the way to maintain them long term, but
this keeps them out of the hair of people working on trunk while
they’re experimental.

Currently only has metainformation and a script to produce a report on
how a CI environment is set up (vital for debugging).

BuildCommon.pm was originally written by me for
libxcrypt (https://github.com/besser82/libxcrypt).  That project uses
very permissive licensing when possible, so it is under CC0 rather
than GPL.