testsuite: log version of shell and make as well
A lot of bugs recently have been due to unexpected changes in shell
and/or make behavior, so let’s log their versions in testsuite.log
to make it easier to figure out when that’s the problem.

Some shells don’t recognize --version but we still get an indication
in the logfile of which shell it is, e.g.

$ make check TESTSUITEFLAGS=1 SHELL=/bin/dash; grep dash tests/testsuite.log
../../tests/local.at:25: /bin/dash --version
/bin/dash: 0: Illegal option --

* tests/local.at: Add ${CONFIG_SHELL-$SHELL} and $MAKE to AT_TESTED.